[Server] Redesign the server configuration
The server persistent state now contains more informations other than the users.
Add a GameDescriptor class to represent the stripped state of a completed game,
that means it holds all the data needed to share the result of the game.
Move out of ClientSession anything related to the user session, that instead
belongs to the class UserSession.